+++ src/Drivers/Gd/Modifiers/ContainDownModifier.php
@@ -2,18 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Intervention\Image\Modifiers;
+namespace Intervention\Image\Drivers\Gd\Modifiers;
 
-use Intervention\Image\Exceptions\InvalidArgumentException;
 use Intervention\Image\Interfaces\ImageInterface;
 use Intervention\Image\Interfaces\SizeInterface;
 
-class PadModifier extends ContainModifier
+class ContainDownModifier extends ContainModifier
 {
     /**
-     * Calculate crop size of the pad resizing process.
-     *
-     * @throws InvalidArgumentException
+     * Calculate crop size of the contain down resizing process.
      */
     protected function cropSize(ImageInterface $image): SizeInterface
     {