+++ changelog.md
@@ -22,6 +22,7 @@
 - ImageInterface::blendTransparency() was renamed to ImageInterface::fillTransparentAreas() - Signature changed & allowed (semi) transparent colors
 - ImageInterface::setBlendingColor() was renamed to ImageInterface::setBackgroundColor()
 - ImageInterface::blendingColor() was renamed to ImageInterface::backgroundColor()
+- ImageInterface::pad() was renamed to ImageInterface::containDown()
 - Config::class option blendingColor was renamed to backgroundColor
 - BlendTransparencyModifer::class was renamed to FillTransparentAreasModifier::class
 - Changed default value for `background` to `null` in ImageInterface::rotate()